Remove --max argument
Instead we automatically stop printing calendar items after 100ms. This
is plenty of time to print a lot of occurrences, more than the user can
reasonably deal with, and should they hit the deadline they'll likely
add an extra filter to bring that number down.

The intent is for this to be the better user experience. 100ms is fast
enough to feel instantaneous, and there's no 'max' setting floating
around that can be set either to low (not enough output) or to high
(slow command).