Remove limits and debugging parameters
I think I'm okay with agenda.txt producing a long-running output if no
filters are specified. It behaves nicely when killed using Ctrl+C.

The one remaining debug flag I'm turning into an environment variable,
to allow setting it in tests and having it available from the start of
the program.